refactor ballot adjudication reasons
- move from utils to interpreter
- change input away from a function
- rewrite testing accordingly
